requirements.txt
glfw
PyOpenGL
numpy
imgui
trimesh
groq
shapely
mapbox-earcut